Update `wodenpy` sky model testing code to create FITS sky models with new polarisation columns and values
